MOD #160 : 검색 로직 ProductReadService로 책임 분리, ProductImageService로 이미지 맵 메서드 분리
