+++ src/main/java/com/lokoko/domain/product/application/service/ProductService.java
@@ -1,24 +1,12 @@
 package com.lokoko.domain.product.application.service;
 
-import static java.util.stream.Collectors.collectingAndThen;
-import static java.util.stream.Collectors.groupingBy;
-import static java.util.stream.Collectors.mapping;
-import static java.util.stream.Collectors.toList;
-
-import com.lokoko.domain.image.entity.ProductImage;
-import com.lokoko.domain.image.repository.ProductImageRepository;
 import com.lokoko.domain.like.repository.ProductLikeRepository;
-import com.lokoko.domain.product.api.dto.response.ProductBasicResponse;
 import com.lokoko.domain.product.api.dto.response.ProductListItemResponse;
 import com.lokoko.domain.product.api.dto.response.ProductStatsResponse;
-import com.lokoko.domain.product.api.dto.response.SearchProductsResponse;
 import com.lokoko.domain.product.domain.entity.Product;
-import com.lokoko.domain.product.domain.repository.ProductRepository;
 import com.lokoko.domain.product.mapper.ProductMapper;
 import com.lokoko.domain.review.dto.request.RatingCount;
 import com.lokoko.domain.review.repository.ReviewRepository;
-import com.lokoko.global.common.response.PageableResponse;
-import com.lokoko.global.kuromoji.service.KuromojiService;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -27,7 +15,6 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.SliceImpl;
 import org.springframework.stereotype.Service;
@@ -39,24 +26,13 @@ import org.springframework.transaction.annotation.Transactional;
 @Slf4j
 public class ProductService {
 
-    private final ProductRepository productRepository;
-    private final ProductImageRepository productImageRepository;
     private final ProductLikeRepository productLikeRepository;
     private final ReviewRepository reviewRepository;
 
-    private final KuromojiService kuromojiService;
+    private final ProductImageService productImageService;
     private final ProductStatsCalculatorService productStatsCalculatorService;
 
     private final ProductMapper productMapper;
-
-    public SearchProductsResponse search(String keyword, int page, int size, Long userId) {
-        List<String> tokens = kuromojiService.tokenize(keyword);
-        Slice<Product> slice = productRepository.searchByTokens(tokens, PageRequest.of(page, size));
-
-        List<ProductListItemResponse> products = buildMainImageResponsesWithReviewData(slice.getContent(), userId);
-
-        return productMapper.toNameBrandProductResponse(products, keyword, PageableResponse.of(slice));
-    }
 
     public Slice<ProductListItemResponse> buildMainImageResponseSliceWithReviewData(
             Slice<Product> slice, Long userId
@@ -73,9 +49,7 @@ public class ProductService {
                 .map(Product::getId)
                 .toList();
 
-        Map<Long, String> imageMap = createProductImageMap(
-                productImageRepository.findByProductIdIn(productIds)
-        );
+        Map<Long, String> imageMap = productImageService.mapMainImageUrlsByProductIds(productIds);
 
         List<RatingCount> stats = reviewRepository.countByProductIdsAndRating(productIds);
         Map<Long, Long> reviewCountMap = productStatsCalculatorService.calculateReviewCount(stats);
@@ -97,25 +71,6 @@ public class ProductService {
         return productMapper.toProductListItemResponses(products, summaryMap, likedIds);
     }
 
-    public List<ProductBasicResponse> makeProductResponse(
-            List<Product> products, Map<Long, ProductStatsResponse> summaryMap, Long userId
-    ) {
-        Set<Long> likedIds = productLikeRepository.findAllByUserId(userId).stream()
-                .map(pl -> pl.getProduct().getId())
-                .collect(Collectors.toSet());
-
-        return products.stream()
-                .map(product -> {
-                    ProductStatsResponse summary = summaryMap.getOrDefault(
-                            product.getId(),
-                            new ProductStatsResponse("", 0L, 0.0)
-                    );
-                    boolean isLiked = likedIds.contains(product.getId());
-                    return ProductBasicResponse.of(product, summary, isLiked);
-                })
-                .toList();
-    }
-
     public Map<Long, ProductStatsResponse> createProductSummaryMap(
             List<Product> products,
             Map<Long, String> productIdToImageUrl,
@@ -131,25 +86,5 @@ public class ProductService {
             summaryMap.put(productId, new ProductStatsResponse(imageUrl, reviewCnt, avg));
         }
         return summaryMap;
-    }
-
-    public Map<Long, String> createProductImageMap(List<ProductImage> images) {
-        return images.stream().collect(groupingBy(
-                img -> img.getProduct().getId(),
-                collectingAndThen(toList(), list ->
-                        list.stream()
-                                .filter(ProductImage::isMain)
-                                .findFirst()
-                                .orElse(list.get(0))
-                                .getUrl()
-                )
-        ));
-    }
-
-    public Map<Long, List<String>> createProductImageUrlsMap(List<ProductImage> images) {
-        return images.stream().collect(groupingBy(
-                img -> img.getProduct().getId(),
-                mapping(ProductImage::getUrl, toList())
-        ));
     }
 }
